fixed invalid  next responder check if not found
A bug exists when indexPathForNextResponderInSectionIndex: on RETableViewCell.m if the sections do not match. If not, the value is set to -1. The for loop below automaticall adds 1 to the unsigned -1, which makes it 0. Thus, the for loop condition was met when it shouldn't have been.

By setting it to -2 then subtracting the number of items in the section, the for loop conidition will never be met when the section is not equal.